requirements.txt
langchain_community==0.0.20
python-dotenv==1.0.1
fastapi==0.109.2
transformers==4.37.2
torch==2.2.0